requirements.txt
fastapi
uvicorn
requests
python-dotenv
sqlalchemy
psycopg2
